+++ Changelog.md
@@ -16,6 +16,7 @@
 - Updated Section model associations with appropriate dependent options to handle cascade deletion while preventing deletion when students exist (#7681)
 - Prevent grade change in the grades table for a Marks Spreadsheet, when scrolling up or down with mouse or keys (#7680)
 - Fixed ordering of sections in the Timed Assessment assignments section-specific settings table (#7690)
+- Fix name column search in graders table (#7693)
 
 ### 🔧 Internal changes
 - Updated Github Actions CI to use cache-apt-pkgs to speed up workflow runs (#7645)